Remove network/replaced with HDD ID
Before with IPv4 it was pulling the lan ip, which, could be change
really easy, thus breaking the auth. I've replaced it with pulling from
a HDD serial. I might think about added a WAN address auth but if the
user moves outside of their home network this would break the auth.